Share the gang_tactics_cards select and row mapper
The select string and embed flattener were duplicated verbatim between
the cached read and the add action. Both now come from types/tactics-card,
alongside withEditionSlug's precedent in types/edition.
